garbage code: now compilation friendly and gofmtd

+++ twistr/card.go
@@ -1,9 +1,9 @@
 package twistr
 
 type Card struct {
-    Aff Aff
-    Ops uint8
-    Name string
-    Text string
-    Star bool
+	Aff  Aff
+	Ops  int
+	Name string
+	Text string
+	Star bool
 }
